+++ circular-buffer.js
@@ -46,11 +46,32 @@ CircularBuffer.prototype.clear = function() {
  */
 CircularBuffer.prototype.push = function(item) {
   if (this.size === this.capacity)
-    throw new Error('mnemonist/circular-buffer: buffer capacity (' + this.capacity + ') exceeded!');
+    throw new Error('mnemonist/circular-buffer.push: buffer capacity (' + this.capacity + ') exceeded!');
 
   var index = (this.start + this.size) % this.capacity;
 
   this.items[index] = item;
+
+  return ++this.size;
+};
+
+/**
+ * Method used to prepend a value to the buffer.
+ *
+ * @param  {any}    item - Item to prepend.
+ * @return {number}      - Returns the new size of the buffer.
+ */
+CircularBuffer.prototype.unshift = function(item) {
+  if (this.size === this.capacity)
+    throw new Error('mnemonist/circular-buffer.unshift: buffer capacity (' + this.capacity + ') exceeded!');
+
+  var index = this.start - 1;
+
+  if (this.start === 0)
+    index = this.capacity - 1;
+
+  this.items[index] = item;
+  this.start = index;
 
   return ++this.size;
 };
@@ -64,7 +85,11 @@ CircularBuffer.prototype.pop = function() {
   if (this.size === 0)
     return;
 
-  return this.items[--this.size];
+  const index = (this.start + this.size - 1) % this.capacity;
+
+  this.size--;
+
+  return this.items[index];
 };
 
 /**
@@ -166,6 +191,13 @@ CircularBuffer.prototype.forEach = function(callback, scope) {
  */
 // TODO: optional array class as argument?
 CircularBuffer.prototype.toArray = function() {
+
+  // Optimization
+  var offset = this.start + this.size;
+
+  if (offset < this.capacity)
+    return this.items.slice(this.start, offset);
+
   var array = new this.ArrayClass(this.size),
       c = this.capacity,
       l = this.size,
@@ -255,7 +287,6 @@ CircularBuffer.prototype.entries = function() {
 if (typeof Symbol !== 'undefined')
   CircularBuffer.prototype[Symbol.iterator] = CircularBuffer.prototype.values;
 
-
 /**
  * Convenience known methods.
  */
@@ -287,7 +318,6 @@ if (typeof Symbol !== 'undefined')
  * @return {FiniteStack}
  */
 CircularBuffer.from = function(iterable, ArrayClass, capacity) {
-
   if (arguments.length < 3) {
     capacity = iterables.guessLength(iterable);